ia/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.4.2
scipy==1.13.0
python-dotenv==1.0.1
psutil==5.9.8